requirement.txt
asgiref==3.4.1
Django==3.2.20
django-ratelimit==3.0.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1